Keep LanguageSelector lighter
